Migrate nameplate to inferno
